refactor: [M3-6318] - MUI v5 - Components > DeletionDialog
