+++ packages/manager/src/components/DeletionDialog/DeletionDialog.stories.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import DeletionDialog from 'src/components/DeletionDialog';
+import { DeletionDialog } from 'src/components/DeletionDialog/DeletionDialog';
 import { action } from '@storybook/addon-actions';
 import type { Meta, StoryObj } from '@storybook/react';
