--- packages/manager/src/components/DeletionDialog/index.tsx
@@ -1,2 +1,0 @@
-import DeletionDialog from './DeletionDialog';
-export default DeletionDialog;
